refactor: move dependency setup to a new file
