+++ internal/server/web_server.go
@@ -7,14 +7,9 @@ import (
 	"net/http"
 	"time"
 
-	"github.com/satsuma-data/node-gateway/internal/checks"
-	"github.com/satsuma-data/node-gateway/internal/client"
-	"github.com/satsuma-data/node-gateway/internal/metadata"
-
 	conf "github.com/satsuma-data/node-gateway/internal/config"
 	"github.com/satsuma-data/node-gateway/internal/jsonrpc"
 
-	"github.com/satsuma-data/node-gateway/internal/metrics"
 	"github.com/satsuma-data/node-gateway/internal/route"
 	"github.com/satsuma-data/node-gateway/internal/util"
 	"go.uber.org/zap"
@@ -30,12 +25,6 @@ type RPCServer struct {
 	httpServer *http.Server
 	routers    []route.Router
 	config     conf.Config
-}
-
-type SingleChainDependencyContainer struct {
-	ChainName string
-	Router    route.Router
-	handler   *RPCHandler
 }
 
 func NewRPCServer(config conf.Config, rootLogger *zap.Logger) RPCServer {
@@ -64,71 +53,6 @@ func NewRPCServer(config conf.Config, rootLogger *zap.Logger) RPCServer {
 	}
 
 	return *rpcServer
-}
-
-func wireDependenciesForAllChains(
-	config conf.Config,
-	rootLogger *zap.Logger,
-) DependencyContainer {
-	var singleChainDependencies []SingleChainDependencyContainer
-	for chainIndex := range config.Chains {
-		currentChainConfig := &config.Chains[chainIndex]
-		childLogger := rootLogger.With(zap.String("chainName", currentChainConfig.ChainName))
-
-		dependencyContainer := wireSingleChainDependencies(currentChainConfig, childLogger)
-		singleChainDependencies = append(singleChainDependencies, dependencyContainer)
-	}
-
-	healthCheckHandler := &HealthCheckHandler{
-		singleChainDependencies: singleChainDependencies,
-		logger:                  rootLogger,
-	}
-
-	mux := http.NewServeMux()
-	mux.Handle("/health", healthCheckHandler)
-
-	for _, container := range singleChainDependencies {
-		mux.Handle(container.handler.path, container.handler)
-	}
-
-	return DependencyContainer{
-		singleChainDependencies: singleChainDependencies,
-		handler:                 mux,
-	}
-}
-
-type DependencyContainer struct {
-	singleChainDependencies []SingleChainDependencyContainer
-	handler                 *http.ServeMux
-}
-
-func wireSingleChainDependencies(config *conf.SingleChainConfig, logger *zap.Logger) SingleChainDependencyContainer {
-	metricContainer := metrics.NewContainer(config.ChainName)
-	chainMetadataStore := metadata.NewChainMetadataStore()
-	ticker := time.NewTicker(checks.PeriodicHealthCheckInterval)
-	healthCheckManager := checks.NewHealthCheckManager(client.NewEthClient, config.Upstreams, chainMetadataStore, ticker, metricContainer, logger)
-
-	enabledNodeFilters := []route.NodeFilterType{route.Healthy, route.MaxHeightForGroup, route.SimpleStateOrTracePresent, route.NearGlobalMaxHeight}
-	nodeFilter := route.CreateNodeFilter(enabledNodeFilters, healthCheckManager, chainMetadataStore, logger, &config.Routing)
-	routingStrategy := route.FilteringRoutingStrategy{
-		NodeFilter:      nodeFilter,
-		BackingStrategy: route.NewPriorityRoundRobinStrategy(logger),
-		Logger:          logger,
-	}
-
-	router := route.NewRouter(config.Upstreams, config.Groups, chainMetadataStore, healthCheckManager, &routingStrategy, metricContainer, logger)
-
-	handler := &RPCHandler{
-		path:   "/" + config.ChainName,
-		router: router,
-		logger: logger,
-	}
-
-	return SingleChainDependencyContainer{
-		ChainName: config.ChainName,
-		Router:    router,
-		handler:   handler,
-	}
 }
 
 func (s *RPCServer) Start() error {
